requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-groq
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search